move faraday client configuration spec set up into shared helper
