+++ spec/lib/common/client/base_spec.rb
@@ -1,16 +1,15 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
+require './spec/support/default_configuration_helper'
 
 describe Common::Client::Base do
   module Specs
     module Common
       module Client
-        class TestConfiguration < ::Common::Client::Configuration::REST
-          def connection
-            @conn ||= Faraday.new('http://example.com') do |faraday|
-              faraday.adapter :httpclient
-            end
+        class TestConfiguration < DefaultConfiguration
+          def adapter_only
+            true
           end
         end
